継承関係のあるクラスをパッケージにまとめた、ViewとModelを繋げた (データやり取りは未検証)
