+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/Bot/BotModel.java
@@ -1,0 +1,7 @@
+package jp.ac.chiba_fjb.app.koutei_chan.Model.Bot;
+
+import jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel.MessageModel;
+
+abstract public class BotModel {
+    abstract public MessageModel reply();
+}
