+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/Bot/GazouChan.java
@@ -1,0 +1,12 @@
+package jp.ac.chiba_fjb.app.koutei_chan.Model.Bot;
+
+import jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel.ImageMessage;
+import jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel.MessageModel;
+
+public class GazouChan extends BotModel {
+
+    @Override
+    public MessageModel reply() {
+        return new ImageMessage();
+    }
+}
