+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/MessageModel/ImageMessage.java
@@ -1,0 +1,7 @@
+package jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel;
+
+public class ImageMessage extends MessageModel {
+    public ImageMessage() {
+        super();
+    }
+}
