+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/MessageModel/MessageModel.java
@@ -1,0 +1,6 @@
+package jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel;
+
+
+abstract public class MessageModel {
+}
+
