+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/MessageModel/TextMessage.java
@@ -1,0 +1,10 @@
+package jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel;
+
+public class TextMessage extends MessageModel {
+    private String message;
+
+    public TextMessage(String message) {
+        super();
+        this.message = message;
+    }
+}
