+++ app/src/main/java/jp/ac/chiba_fjb/app/koutei_chan/Model/ModeModel.java
@@ -1,8 +1,8 @@
 package jp.ac.chiba_fjb.app.koutei_chan.Model;
 
-import jp.ac.chiba_fjb.app.koutei_chan.Model.BotModel;
-import jp.ac.chiba_fjb.app.koutei_chan.Model.KouteiChan;
-import jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel;
+import jp.ac.chiba_fjb.app.koutei_chan.Model.Bot.BotModel;
+import jp.ac.chiba_fjb.app.koutei_chan.Model.Bot.KouteiChan;
+import jp.ac.chiba_fjb.app.koutei_chan.Model.MessageModel.MessageModel;
 
 public class ModeModel {
     private boolean questionFlag = false;
